requirements.txt
aiohttp==3.6.2
appnope==0.1.0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.4.5.2
cffi==1.14.3
chardet==3.0.4
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
idna==2.9
importlib-metadata==4.8.1
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab==2.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
kiwisolver==1.3.1
llvmlite==0.33.0
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
multidict==4.7.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numba==0.50.0
numpy==1.18.5
oauthlib==3.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
plotly==4.14.3
prometheus-client==0.9.0
prompt-toolkit==3.0.8
psycopg2-binary==2.9.1
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-rapidjson==0.9.1
pytz==2020.1
pyzmq==20.0.0
rapidjson==1.0.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
tabulate==0.8.9
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.2
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
yarl==1.4.2
zipp==3.5.0
zope.interface==5.1.0
